#39 Rename domain as entities and fix imports
