+++ webminer/entities/arxiv_document.py
